Cherry pick h22o (#786)
* Http2 to origin

This implements the HTTP/2 to origin feature.

* Adding window size and flow control out parameters.

* Handle set_connect_fail in ConnectingEntry correctly

When ConnectingEntry handled a connection failure, it could clear the
error value passed to set_connect_fail. This would incorrectly result in
the server not being marked down by hostdb. This patch ensures that the
error code passed to set_connect_fail is always an error value when
handling a connection failure.

* Send WINDOW_UPDATE for session when stream ends for other streams

* pass through RST_STREAM frame

* set_cause_of_death_errno

* Http2 to origin fix (#4)

* Fix for connection level window mismatch + force connection level window update

* Fix for HPACK corruption on closed streams
Todo: check stream level error returns, these should still decode
  the headers to prevent HPACK corruption for following header
  frames

---------

Co-authored-by: Kees Spoelstra <kd.spoelstra@gmail.com>

* proxy.config.exec_thread.autoconfig -> ''.enabled

This is a changed needed to the new HTTP/2 to origin tests for the
records.yaml update.

* Revert a restart_receiving call causing POST issues

The post-continue.test.py AuTest demonstrated that the restart_receiving
call was causing the first 64KB of data from a POST to be lost.

* Expand tunnel_handler to expect VC_EVENT_INACTIVITY_TIMEOUT

The HttpSM::tunnel_handler expects the tunnel to be going away and
asserts that the event it receives is HTTP_TUNNEL_EVENT_DONE. However,
the tunnel can be going away due to timeout as well. Adding
VC_EVENT_INACTIVITY_TIMEOUT as a possible event in that function.

* remove the buffer reader from the consumer's vc

Co-authored-by: Brian Neradt <brian.neradt@gmail.com>
Co-authored-by: Fei Deng <feid@oath.com>
Co-authored-by: keesspoelstra <kd.spoelstra+github@gmail.com>
Co-authored-by: Kees Spoelstra <kd.spoelstra@gmail.com>