Fixing remap acl rules for @action=allow
If a user specifies an @action=allow remap.config ACL rule, then the
implication is that requests with methods not in the allow list would be
denied. Before this patch, allow ACL rules would just never deny. This
fixes that behavior so that allow ACL rules that match on IP but not on
method deny.