Fix an infinite loop in Histogram (#11752) (#984)
The loop in Histogram::operator() is infinite if R = 7, S = 2, and sample = 512.  When this happens, the histogram will not find a matching bit and will loop forever with mask at 0.  This change lowers OVERFLOW_BOUND so that every sample is guaranteed to have a 1-bit within the range of histogram buckets.

(cherry picked from commit 36da45475db02790b95d615a690b91170aa2f06c)

Co-authored-by: Mo Chen <mochen@apache.org>