test update to reproduce kem logging issue
Build openssl 3.5:

```
sudo ./tools/build_openssl_3_5.sh
```

Then generate the new kem certs:

```
bash tools/generate_kem_cert.bash
```

Those new certs are copied over server.* and signer.* for the ssl test
in this patch.

Then run the test to reproce the issue:

./autest.sh --sandbox /tmp/sb --clean=none -f tls_client_verify

╰─➤  cat /tmp/sb/tls_client_verify/ts/log/squid.log
- 404 http://127.0.0.1:61006/case3 0 0

Note the leading `-` - that is the reproduction of the issue.