Plugin leak fixes (#1078)
* ja3_fingerprint: fix leak of duplicate header fields

When looping through duplicate header fields with TSMimeHdrFieldNextDup,
each iteration must free the TSMLoc. ja3_fingerprint had a loop where it
did not do this, thus it leaked these. This adds the required
TSHandleMLocRelease.

* xdebug: release the client request MLoc

XScanRequestHeaders missed a TSHandleMLocRelease call for the client
request MLoc resource.

* stats_over_http: ensure BrotliEncoderDestroyInstance

stats_over_http calls BrotliEncoderCreateInstance, but we weren't always
calling BrotliEncoderDestroyInstance. This adds it to the b_stream
destructor to ensure it gets called.