Reuse prepared cache write on redirected request (#11542) (#1120)
* Add AuTest to reproduce #9275

This reproduces a variant of #9275 where we fail on the cache lookup. If
the cache lookup were to succeed, it should also fail on the cache write,
so this one test should cover both bugs.

* Reuse prepared cache write on redirected request

Fixes #9275.

When there is no parent proxy and ATS is following a redirect, it will do
a cache lookup on the redirected request. If that lookup results in a cache
miss, it will start to prepare a new cache write, which is wrong. The state
machine intentionally leaves the write open so we can re-use the connection;
this is possible to ascertain from the code and from comments to that effect.

We should only open a new write if we don't already have one when following
a redirect.

(cherry picked from commit 875463c9645787d380b3215b414e906c6c38cd7b)

Co-authored-by: JosiahWI <41302989+JosiahWI@users.noreply.github.com>